requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-auth
